+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## pre-release v0.5
 
 - Removed **xISHServer** module from this repository and moved to [ISHServer](https://github.com/Sarafian/ISHServer).
+- Fixed some issues with the `Restart-Server.ps1` and `Test-Server.ps1` scripts.
 
 ## pre-release v0.4